requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
dlib==19.22.1
scipy==1.7.1
Pillow==8.3.1
